Refactor schemaGenerator.ts to improve code readability and remove unnecessary imports
