+++ src/schemaGenerator.ts
@@ -1,21 +1,12 @@
-// ! if prisma.schema exists, then appends after last model
-// ? Partially implemented TODO: add secondary parser, to test for any possible bugs and recommend solutions, like automincement: true, type: int only.. so on..
-// DONE TODO: if foreign key is false, type != anyModelType
-
-// TODO: Add comment that this model is generated by schemaGenerator
-// TOdo: create issue, if same model name defined in schema, then it should throw error in console, and not generate prisma schema
-
 import { createModel, createScalarField } from "prisma-schema-dsl";
-import {
-  AUTO_INCREMENT,
-  DataSourceProvider,
-  DataSourceURLEnv,
-  Enum,
-  ScalarType,
-  UUID,
-} from "prisma-schema-dsl-types";
-import { Field, Schema } from "./dynamoPrisma.types";
+import { AUTO_INCREMENT, ScalarType, UUID } from "prisma-schema-dsl-types";
+import { Field, Schema } from "./types.dynamoPrisma";
 
+/**
+ * Creates models based on the provided schema.
+ * @param schema The schema object containing the schema definition.
+ * @returns An array of model objects.
+ */
 export function createModels(schema: Schema["schema"]): any[] {
   const models: any[] = [];
   for (const schemaItem of schema) {
@@ -25,34 +16,21 @@ export function createModels(schema: Schema["schema"]): any[] {
   return models;
 }
 
-// increment is breaking the code
-// ! ERROR: Error parsing JSON: Error: Default must be a number or call expression to autoincrement()
-// ? Log by console.war:
-/*  String;
-    Default
-    Default
-    Default
-    Default
-    Default
-    Int 
-*/
+/**
+ * Creates an array of fields based on the provided field data.
+ * @param fields - An array of field data.
+ * @returns An array of fields.
+ */
 export function createFields(fields: Field[]): any[] {
-  // console.error("Feilds: ", fields);
   const result: any[] = [];
   for (const fieldData of fields) {
     fieldData.isId && fieldData.autoincrement
       ? console.error("Cannot have String in autoincrement")
       : null;
 
-    // TODO @db.Uuid if default is uuid, and isId is true
     result.push(
       createScalarField(
         fieldData.fieldName,
-        // fieldData.type as ScalarType
-        //
-        //   : fieldData.isId && fieldData.uuid
-        //   ? ScalarType.String
-        //   :
         fieldData.type as ScalarType,
         fieldData.isList || undefined, //isList boolean | undefined
         !fieldData.nullable || false, //isRequired boolean | undefined
@@ -104,6 +82,5 @@ export function createFields(fields: Field[]): any[] {
       );
     }
   }
-  // console.log("Results: ", result);
   return result;
 }
