+++ src/types.dynamoPrisma.ts
